\name{d5:lymphoma}

\docType{data}

\alias{lymphoma}

\alias{lymph.X}

\alias{lymph.y}


\title{Lymphoma Microarray Data}

\description{
This is one of the microarray data sets used to demonstrate BCBCSF in the reference article. Information about this data set can be found from the reference. 
}

\usage{
    data(lymphoma)
}

\value{
\item{lymph.X}{a matrix of gene expression data for p = 4026 genes on n = 62 cases in G=3 classes}

\item{lymph.y}{a vector of class labels coded by 1,2,3.}
}


