\name{internal}

\title{Internal functions used in package `gausspred'}

\alias{comp_amlp}

\alias{comp_eer}
 
\alias{comp_loss}

\alias{gen_qflmd}

\alias{mcmc_pred}

\alias{met_gauss}

\alias{mk_folds} 

\alias{mlepred}

\alias{probs_attrue}

\alias{rank_F}

\alias{richisq}

\alias{rinvgam}

\alias{trpr_mle}

\alias{log_sum_exp}

\alias{comp_adjfactor}

\description{This function is internal. Type the function name to look at its details}

\keyword{internal}


